# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Integration Swagger in Laravel with Passport Auth Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/personas": {
            "post": {
                "tags": [
                    "Persona"
                ],
                "summary": "Registro de personas",
                "description": "Acá se registran las personas familiares de usuarios",
                "operationId": "Crear Persona",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cedula",
                                    "nombre1",
                                    "nombre2",
                                    "apellido1",
                                    "apellido2",
                                    "fecha_Nacimiento",
                                    "sexo_id"
                                ],
                                "properties": {
                                    "cedula": {
                                        "type": "text"
                                    },
                                    "nombre1": {
                                        "type": "text"
                                    },
                                    "nombre2": {
                                        "type": "text"
                                    },
                                    "apellido1": {
                                        "type": "text"
                                    },
                                    "apellido2": {
                                        "type": "text"
                                    },
                                    "fecha_Nacimiento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "sexo_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Registrar"
                ],
                "summary": "Registro de usuario",
                "description": "Acá se registran los usuarios de la UCR sede Guanacaste",
                "operationId": "Register",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "cedula",
                                    "nombre1",
                                    "nombre2",
                                    "apellido1",
                                    "apellido2",
                                    "fecha_Nacimiento",
                                    "sexo_id",
                                    "email",
                                    "password_",
                                    "c_password"
                                ],
                                "properties": {
                                    "cedula": {
                                        "type": "text"
                                    },
                                    "nombre1": {
                                        "type": "text"
                                    },
                                    "nombre2": {
                                        "type": "text"
                                    },
                                    "apellido1": {
                                        "type": "text"
                                    },
                                    "apellido2": {
                                        "type": "text"
                                    },
                                    "fecha_Nacimiento": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "sexo_id": {
                                        "type": "integer"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "test@ucr.ac.cr.com"
                                    },
                                    "password_": {
                                        "type": "password"
                                    },
                                    "c_password": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Register Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login de usuario",
                "description": "Acá se loguea el usuario de la UCR sede Guanacaste",
                "operationId": "authLogin",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password_"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "email",
                                        "example": "test@ucr.ac.cr.com"
                                    },
                                    "password_": {
                                        "type": "password"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "200": {
                        "description": "Login Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    }
}